feat: Sound에 카테고리 필드 추가 #11
